Move classes `RulesDumper`, `RuleHandlerContainer` and `SimpleRuleHandlerContainer`  (#454)
